Fix release cut off check
Previously the cut off warning would always appear if the current hour
was 14 or higher. This meant that a release two or more days into the
future could show the warning prematurely (e.g. we should be able to add
tickets to a release on Wednesday after 2pm on Monday but the warning
was still showing).

Now, the cut off warning actually takes into account the date of the
release and only shows the warning after 2pm on the last working day
before the release.