Allow audit URLs to be used instead of diffs
Sometimes, we need to release quick fixes that were not reviewed. We
still want to add these to the release app for visibility and, rather
than linking to nothing, we can link to the audit on phab.

For example, link to
http://phab.dev.ebuyer.com/rGENONEfcc395703a1f8edee04db518ed56c121705900fe
rather than http://phab.dev.ebuyer.com/D5610

This patch updates the Add Diff form to indicate this possibility, and
improves the styling of the release view so that lengthy commit hashes
do not overflow other elements of a release (such as the Release
button/Released status text).